Invalid left-hand side in assignment

+++ events/messageCreate.js
@@ -25,15 +25,17 @@ module.exports = {
                 break;
             };
             case `966063122584588378`: {    // digital's shit, k thx
-                if (!message.member?.user.bot && !message.webhookId && (message.member?.id != '797012765352001557' || message.member?.id != '821472922140803112')) {
+                if (!message.member?.user.bot && !message.webhookId && (message.member?.id != '797012765352001557' && message.member?.id != '821472922140803112')) {
                     message.member.ban({ reason: 'Wrote in disalowed channel' }).catch(() => null);
                     message.delete().catch(() => null);
                     return log(`App`, `Beanned ${message.member.user.tag}`, `green`);
                 };
                 if (!message.content && !message.embeds) return;
-                message.embeds[0]?.color = `#7289da`; message.embeds[0]?.author?.name?.replace('Crni39', 'The Digital');
-                message.embeds[0]?.description = '> ' + message.embeds[0]?.description?.split('- Crni39\n').join('\n> ').slice(0, -2);
-                
+                if (message.embeds[0]) {
+                    message.embeds[0].color = `#7289da`; message.embeds[0]?.author?.name?.replace('Crni39', 'The Digital');
+                    message.embeds[0].description = '> ' + message.embeds[0]?.description?.split('- Crni39\n').join('\n> ').slice(0, -2);
+                };
+
                 await require('axios')({
                     url: client.webhook,
                     method: 'POST',
